+++ apps/blueprints/__init__.py
@@ -1,0 +1,6 @@
+# -*- encoding: utf-8 -*-
+"""
+RijanAuth - Blueprints Package
+"""
+
+# Blueprint imports will be registered here as they are created